requirements.txt
requests==2.30.0
environs==9.5.0
python-telegram-bot==13.15
redis==4.5.5
